Added changes in the file structure
