+++ src/pages/Register/Register.js
@@ -1,29 +1,34 @@
 import React from 'react';
-import './Login.css';
+import '../Login/Login.css';
 import { CardImg, FormGroup, Label, Input, Button } from 'reactstrap';
-import swinai from './swinai.png';
 import { useNavigate } from 'react-router-dom';
+import swinai from '../../assets/swinai.png';
 
-const Login = () => {
+const Register = () => {
     const navigate = useNavigate();
-    
+
     return (
         <div className="background">
           <div className="login-box">
             <div className="container">
-              <div class="row app-des">
-                <div class="col left-background ">
-                 
+              <div className="row app-des">
+                <div className="col left-background">
                   <CardImg
                     className="mobile-img"
                     src={swinai}
                     alt="mobile-App"
                   />
                 </div>
-                <div class="col login-form">
+                <div className="col login-form">
                   <form>
-                    <h2 className="font-weight-bold mb-4">Login</h2>
+                    <h2 className="font-weight-bold mb-4">Register</h2>
                     <FormGroup>
+                      <Label className="font-weight-bold mb-2">Full Name</Label>
+                      <Input
+                        className="mb-3"
+                        type="text"
+                        placeholder="John Doe"
+                      />
                       <Label className="font-weight-bold mb-2">Email</Label>
                       <Input
                         className="mb-3"
@@ -36,23 +41,27 @@ const Login = () => {
                         type="password"
                         placeholder="At least 8 characters"
                       />
+                      <Label className="font-weight-bold mb-2">Confirm Password</Label>
+                      <Input
+                        className="mb-3"
+                        type="password"
+                        placeholder="Confirm your password"
+                      />
                     </FormGroup>
-                    <Button className="mt-3  btn">Login</Button>
+                    <Button className="mt-3 btn">Register</Button>
                     <Button 
-                      className="mt-3  btn"
-                      onClick={() => navigate('/register')}
+                      className="mt-3 btn"
+                      onClick={() => navigate('/login')}
                     >
-                      Don't have an account? Sign up
+                      Already have an account? Login
                     </Button>
-                    {/* <div className="text-center m-4">or continue with social account</div>
-                  <GoogleLoginButton className="mt-3 mb-3 px-auto text-center"/> */}
                   </form>
                 </div>
               </div>
             </div>
           </div>
         </div>
-);
+    );
 };
 
-export default Login;
+export default Register;
